Basic framework for implementing containers and layout managers.
A container is an element, so containerObj inherits from elementObj, ditto
for the implementation objects.

Each container has a layoutmanager, which consists of a public object
and a private implementation object.

Define a stub for the grid layout manager.

Change generic_window, and main_window to be containers. create_mainwindow()
supplies a default layoutmanager, and the real create_mainwindow() takes
a new_layoutmanager parameter.