Change inheritance of the container implementation object.
Do not have the container implementation object inherit from the element
implementation object.

The subclass of containerObj::implObj must implement get_element_impl()
and get_window_handler(), accordingly.

Add a constructor to horizvertObj that takes a pair of axis specifications.

Add a constructor to elementObj::implObj that takes a pair of axis
specifications, to be forwarded to horizvertObj.