Insert elements.
Mutex-protected map in the gridlayoutmanager implementation object stores
child elements that are in the grid, by their x/y coordinates.

insert() inserts them in there, and leaves it at that.

recalculate() calls rebuild_elements(), which takes them out of the map,
and unwinds them into a vector, precalculating the data structure that
grid metrics calculations need.