Framework for theme support.
A config file in the home directory specifies the default theme.

The library checks if the CXXWTHEME property in the root window is set, if
not the theme name is read from the config file, and installed into the
CXXWTHEME property.

The defaulttheme object parses the theme file.

Implement parsing of dimensions and colors.