Link pictformats to the connection they came from.
Store a pointer to the connection of a pictformat in the pictformat
implementation object. We want to make sure that someone doesn't use
a pictformat from one connection with something on another connection.

The sanity check in change_gc_locked() can simply cmpare pointers to the
screen to make sure they're the same.