Add corner border elements to the grid layout manager.
Add the "corner_borders" container.

Rename straight_border_hash_map as border_hash_map, since it can be used
as the hasher for both corner_borders and straight_borders container.

Factor out the update_borders closure in rebuild_elements() into the
get_straight_border() method. Add get_corner_border() to perform
similar functionality for a corner border. After creating a straight
border, the straight border gets connected to the corner border.

In order to be able to refactor this, need to expose the signatures
for the straight border factories and update methods.