Fill in the empty space in rounded corner elements ....
... with the background color from the grid element whose corner border
is rounded.

Turns out that the placeholder bool flags, in various border routines,
need to be grid_lementptrs, for the appropriate border element.

Pass surrounding_elements_info to draw_corner(). Before it draws a
multiple-corner border, it fills in each one of the corners with the
corresponding grid element's background color.

Pass the appropriate corner element to draw_cornerXX(), on its merry way
to draw_round_corner(), where it gets used to fill in the inner gap in the
rounded corner border.

The bulk of the work is in corner_border(), to carefully fill in each
quadrant of the corner border with the corresponding element's background
color.