Support padding of grid elements.
A bunch of moving pieces of the puzzle. Starring container_clear_padding(),
that figures out the padding area of an element in the grid, and
uses clear_to_color() to wipe it out.

To make that happen, clear_to_color() gets an additional parameter,
a second draw_info, for an element whose background color gets used to
clear this element's background. By default, it's the same element,
but when we're clearing the padding area in the container, this is used
to obtain the element's background color.

The whole show gets kicked off in the grid layout manager. padding() sets
a new element's padding, in millimeters. When the grid elements get built,
the padding gets translated using the current theme's scaling factors into
pixels.

The padding dimensions get factored into two place:

- In apply_metrics(). After obtaining the display element's metrics, the
padding gets added in.

- In process_updated_position(), subtracting it from the grid cell's
initial dimensions, before aligning the element, if needed.

Display optimizations:

1. Do not reuse child_background_color_changed() to implement child
visibility updates. Call redraw_child_borders_and_padding() if the child
element is visible.

2. The existing child_background_color_changed() code, that triggers
the redrawing of the surrounding borders, becomes a part of
redraw_child_border_and_padding(), which also takes care of clearing
the padding area.

3. When updating element visibility, replace the bool flag, indicating the
new visibility status, with "inherited_visibility_info", consisting of the
flag, but also a second flag that suppresses the redrawing caused by the
visibility change. If the visibility change was handled in
generic_window_handler via MapWindow(), this flag is used to suppress the
redraws, since it's expected that the forthcoming exposure events will take
care of that.