Refactor construction of richtext.
Makes it possible to pass a richtextstring to richtext's constructor, to
set its initial contents.

Factor out the bits of construction that must be done IN_THREAD_ONLY, and
simply make sure that those bits gets done when the public object locks
the implementation object. The first time the object gets locked, in thread,
the initialization will finish.