Documentation, and an examples directory with helloworld
Rename mcguffin() to connection_mcguffin(), and provide a convenience
connection_mcguffin() in the element base class, so that it's not necessary
to pull in the screen and connection header files in order to obtain the
connection mcguffin.

Turns out that it wasn't needed to have current_fontcollection visible in
the public API, so move the current_fontcollectionfwd.H header file.

Swap the layout manager and the creator lambda parameters to
create_mainwindow(), to be more consistent.

Create an RPM package spec file.

Fix some compiler warnings.