Focus frame container.
A focusable container object, based on the grid layout manager, but which
expects to contain a single display element, installed via set_focusable().

Refactor code that creates theme-based borders. The focus frame container
uses a subclass of the grid layout manager that installs one of the focus
borders around the display element in the container. The focus frame
container overrides focus_event() and triggers a rebuild of the container,
in order to set the focus frame, accordingly.