Word-wrappable labels.
Initial width of word wrapped text gets specified in millimeters.

Change richtextObj::get_metrics() to return that as the preferred width,
and add calculations to the internal richtext object to compute the
minimum width of the paragraph and the entire text. The code already
computed the minimum for each fragment.