Rich text iterators.
richtextiterator objects allow modifications of rich text objects that
will be used to implement input fields. richtextiterator is the public
interface to a richtext, and a richtextcursorlocation. richtextcursorlocation
is really the internal implementation object of richtextiterator.

richtext's begin(), end(), and at() return richtextiterators. next(),
prev() and move() methods move the richtextiterators inside the
richtextitetator. insert() and remove() modify the contents of the richtext.
get() returns the contents of the richtext.