Framework for implementing viewport.
Add ensure_visibility() method to the element implementation object, for
an element to request some part of it to be visible.

This bubbles up to the new peephole layout manager, which makes sure that
the requested portion of the element in the peephole is visible through the
peephole.

The peephole layout manager always contains a single element, that the
peephole layout manager positions inside itself, without regards to the
peephole's dimensions, and the existing drawing framework should clip the
child element's rendering to the peephole.