Text input field.
Factor out editor_container's override of request_visibility_recursive into
a nonrecursive_visibility mixin.

Split the parameters to focusframecontainer's constructor, pass a
separate focusable_impl parameter for constructing the focusableObj::ownerObj
superclass; with actionbutton passing the same implementation object.

An input_field consists of an outer container. The inner container has
a focusframe with an editor_container, with the editor's focusable
implementation object used to construct this focusframe, as the
focusable implementation object.