Refactor conversion of selections.
Remove code from window_handler that deals with conversions of pasted text
to unicode.

What's left in window_handler deals purely with handling property notify
messages, and defining a converted selection API that provides the converted
data to the subclass, dealing with the SelectionNotify and PropertyNotify
events, and converting them to the API calls.

generic_window implements the API and converts the pasted text to unicode,
then punts it to the current element with the input focus.