XIM framework.
Construct a second, INPUT_ONLY window, that will talk to the X Input Method
Server.

Define XIM messages in XML format, with a stylesheet to robo-generate
code to create and parse the messages.

The ximserver handles the details of the protocol implementation.
The ximxtransport implementation object inherits from ximserver, and
takes care of the X transport over which the XIM protocol is implemented.

The generic_window implementation object owns two more objects, a
XIM transport object (the XIM protocol interface), and a XIM client
object, which coordinates the interface between the generic_window, and
everything in it, with the XIM transport object.

The XIM transport object is actually stored in the connection object,
but weakly. It's instantiated when needed; as such the XIM transport
object is a singleton, and the same XIM transport object will be used
for all top level generic_windows.

The XIM transport object owns the XIM server object, which actually handles

Carefully arrange for circular references between generic_window's
window_handler superclass, and the XIM server and client objects.