Link labels to focusable elements.
Add a linking object, label_for() that links a display element to a focusable.

When drawing a display element, draw it with a dithered disabled mask if
the element is linked to a focusable and that focusable is disabled, even
if the display element is supposedly enabled.

A focusable keeps a list of all of its labels, and makes sure to schedule
their redrawing when its enable status changes. This also is checked
in focusable_deinitialize(), as well, for good measure.

Also, when a button is pressed on top of a display element, if it's a label
for a focusable forward the button press to the focusable element too.
In order to do that the element must save the last motion event it received,
so that it can be replayed for the focusable, before forwarding the button
event.