Focus fixes
Button is the focusable element. The label in the button wasn't being
drawn as disabled, since the button's enabled flag controlled focusability.

Have an element's enabled() overridden and used to check its parent container's
enabled. So, if a parent is disabled, the element is considered to be
disabled as well. Always use enabled() instead of checking the enabled flag
directly. enabled() will also check the parent.

When setting a focusable's enabled status, recursively redraw everything in
the focusable.