Add scrollbars to peepholes.
Have editor_container inherit from peepholed_element, an interface for
an element in a peephole, which must specify its nominal scrolling
increment.

When creating the editor container, also create the two scrollbars for
the peephole that will go around the editor. The two scrollbars are also
used to initialize the peephole layout manager, which supplies the
implementation objects that notify the peephole when the scrollbars
are moved.

The input field is responsible for inserting the scrollbars into the
top level container.