Framework for a list display element.
Factory's create_list() returns a focusable_container with a
listlayoutmanager.

Make list_style specify more visual properties of the list.

Factor out the mixin class for the implementation object used for
the input_field's peephole, and use the same implementation object
for the list's peephole, the peepholed_container_impl_t.

Have listcontainerObj inherit from a focusable_container, make the
list container a focusable display element.

The actual list is a peephole with a focusframe. The peephole contains
a listcontainer with listlayoutmanager as its layout manager.

The listcontainer knows that it's inside the peephole, and updates the
peephole's metrics based on its own metrics, and the number of visible
rows specified by the list style.