Reimplement batch recalculation scheduling, redraw optimization.
Move the set of containers queued up for recalculation from the batch_queue
object into the connection thread's execution context.

A batched recalculation gets queued up by sending a message to the connection
thread to add a container to the queue.

The only thing that's left in the actual batch queue object is the
something_scheduled indication, and the batch queue's destructor, if
something_scheduled, sends an execute_batched_jobs() message to the
connection thread to execute_batched_jobs(), which calls needs_recalculation()
on the containers that were scheduled for batch recalculation, then
process_events().

Add a has_exposed flag to generic_window_handler, that's set when an exposure
event is received, and cleared when a configure event is received (with the
expectation of an exposure event in the works). Change schedule_redraw()
to schedule the element to be redrawn only if has_exposed.

request_visibility() will batch the visibility request.