Refactor motion event reporting.
Rename the existing motion_event() handler to report_motion_event(), and
make it possible to discrement real motion event versus fake ones.

Collect the existing parameter to the handler into a new motion_event
class, and include a motion_event_type that specifies whjich kind of
motion event it is.