Automatically parse theme dimensions specified as raw mms.
Move compute_width() and compute_height() from the screen and into the
screen's theme object. Change all calls to them to call them as the
theme's methods.

Replace get_theme_dim_t() with get_theme_width_dim_t() and
get_theme_height_t(). If the given dimension is not found, an attempt is
made to parse it as a double, and if that succeeds use
compute_width() or compute_height().

Change themedimObj to take a pointer to either the get_theme_width_dim_t()
or the get_theme_height_dim_t() class method, and use the appropriate
method for the calculation.