Do not switch input focus to combobox popup.
At least the xfce window manager draws the main window as losing input
focus, even though the popup is tagged as a popup of the application
window.

Instead, redirect the pointer and keyboard events internally.

Also refactor the invocation of the selection changed callback, and
construction of the internal layout manager implementation object.