Optimize setting of WM_SIZE_HINTS.
No need to set them if the main window is not visible.

Overridden horizvert_updated() continues to keep track of the
preferred dimensions, but bail out if the main window is not visible,
or call update_size_hints() when it is visible.

Override set_inherited_visibility and call update_size_hints(), so that
WM_SIZE_HINTS gets set only immediately prior to mapping the window.