Allow text hotspot links in focusable labels.
Optional parameter to richtextstring appends a null byte to the
rich text string. text hotspot links will be tracked by beginning and
ending richtextiterators, which cannot point to the real end(). Their end()
pointers to the last character in the richtextstring, which will be this
null byte.

Refactor construction of editor_impl subclass of textlabel, to be
compatible with the new world order.