Refactor the logic for dim_t::infinite maximum metrics of a grid element.
Instead of skipping everything with maximum() is dim_t::infinite, if it is
then set it the width or the height.

The if() statement takes effect only if the maximum() is less than the
element_position, so when it was originally dim_t::infinite() it gets skipped.

This was a no-op anyway, since infinite() is largest possible value.