Fix metrics calculations.
Fix logic used by wrappable text labels to set their metrics, to avoid
needless recalculations of metrics of all of their containers.

Their preferred width is fixed, and does not get adjusted when the
label element is visible. That's it.

Have richtext cache the word_wrap_width in rewrap, to avoid expensive
recalcs.