Use CopyArea() to efficiently scroll peepholes' content.
When setting the new position of the peepholed element, if its size is
unchanged this must be a simple scroll. If the amount scrolled is
less than the peephole's size, use CopyArea() to scroll the peephole, thus
avoiding the need to redraw the entire peephole from scratch.

Redraw the bare minimum that's left to be redrawn after the scroll.