Do not focus automatically on the first button in the color picker.
Or the first chevron in the date input field's popup.

Remove the call to set_default_focus() when a popup becomes visible.
The dropdown list relies on this to make the dropdown list have keyboard
focus when it's shown. Instead, simply enable autofocus() for the dropdown
list's internal list element.