Fix shortcut activation.
When we're in a popup and press enter, don't activate some button on the
parent display element that uses enter as a shortcut.

Have separate shortcuts installed in generic_window_handlerObj, in
addition to the shortcut collection in shared_handler_dataObj.
The ones in generic_window_handlerObj are local, the ones in
shared_handler_dataObj.

Have handle_key_event search its local shortcuts and the global ones.
Install list item shortcuts in the global list, and all others in the
local list.