Refactor key handling.
The end goal is to turn off the button when release a key which opened
a popup. Because a popup started grabbing key events as soon as it
opened, the hotspot button that opened the popup never sees the key
release events, and never un-highlights itself.

Also fixes the relative handling of XIM. Forward key events to popups
before they're sent to the XIM server. Introduce another layer in the
onion: forward_key_event_to_xim(). Change forward_key_event() to call
shared handler data's handle_key_event() which forwards the event
to forward_key_event_to_xim() in the active popup, and if there's no
popup then forward_key_event() calls its own forward_key_event_to_xim().
forward_key_event_to_xim() then hands the baton to XIM, which returns it
by calling handle_key_event() as usual, where things return to normal.

Meanwhile we add a reporting_key_event_to() hook in shared handler data.
It takes key release events, checks if it goes to a popup, and if so,
looks at all OTHER popups, and the original generic window, and if any
of them have an element with most_recent_keyboard_focus(), that focusable's
new grabbed_key_event() hook is called.

Implement this hook in the hotspot to set the hotspot temperature back to
cold (or maybe warm). This results in the button going dark after it opens
a popup and receives a key release event.
receive that