Optimize grid layout manager forced recalculation.
The grid layout manager forcibly schedules itself for position recalculation
when child metrics change, in order to reposition its child elements as needed.
Optimize this by tracking whether the position of an element really changed,
and if not invoke a new callback process_same_position() instead of
process_updated_position().