Fix height of the combo-box's current value element.
It was based on the height of the elements in the combo-box list, plus
extra padding. Need to make it based on the height of the elements without
the padding.

Split the value that holds the height of the tallest list element into two
values, with/without padding, and use the right value, in each context.