Finish computation of label metrics in textlabel_implObj constructor.
Calling compute_width() to compute the preferred_width in
textlabel_implObj's constructor, and then calling text->rewrap() allows
textlabel_implObj's constructor to terminate with the initial calculate
of the label's metrics being complete.

This is because text->rewrap() no longer requires to be called IN_THREAD
(although if not, its results are immediately meaningless upon return,
that's ok, we will recheck the results when we initialize() the label
in the connection thread).

Also, proactively, mark calculate_current_metrics() as long requiring
IN_THREAD invocation.