Draw an ellipsis to indicate a truncated label.
Refactor draw_with_ellipsis out of draw().

draw() checks if there's a richtext_allipsis, and if so it passes its first
(and only) fragment to draw_with_ellipsis, otherwise it passes a null ptr.

Factor out the inner loop in the former draw(), now draw_with_ellipsis(),
into a helper inner class, draw_fragment_info. This helper class contains
the data the inner loop uses, as class members. Before the loop, instead
of initialize the former function variables, draw_with_ellipsis() initializes
an instance of this draw_fragment_info helper clsas.

Then, each fragment is checked, if its width exceeds the width of the
drawing bounds area, and the ellipsis fragment is provided. If so,
the draw bounds are adjusted even smaller, to make room for the ellipsis,
and the fragment gets drawn as usual, by calling the factored out
code that's now sitting in draw_fragment_info.

Then, the drawing bounds are reset for the ellipsis part, and, in this
manner, draw_fragment_info gets simply recycled for drawing the ellipsis.

Some cleanup in the textlabel object. Since word_wrap_widthmm is const it
doesn't need to be IN_THREAD protected. If the truncatable flag is set,
adjust the horizontal metrics so that they're scalable.