listlayoutmanager's container is always listcontainer_pseudo_implObj
listlayoutmanagerObj::implObj's container is always a
listcontainer_pseudo_implObj, so store the link to it in the
layoutmanager object, accordingly.

Similarly, peepholed_listcontainer_impl is always mixed in with
a listcontainer_pseudo_implObj, so replace get_peepholed_listcontainerObj()
that returns a child_elementObj with get_pseudo_impl().

The peepholed_toplevel_listcontainer_layoutmanager_implObj layout
manager subclasses listlayoutmanager, but takes a
peepholed_toplevel_listcontainer_implObj as it's container, so add an
additional parameter, to avoid dynamic_casts.