Means to update button contents.
Make arrangements for a button's get_layoutmanager() to return the
internal singletonlayoutmanager, making it possible to update
the button's contents.

Singletonlayoutmanager's bugs: make sure it correctly handles
element initialization and removal semantics. Make sure that
it positions the new element whose metrics are the same as the old
one's. recalculate() needs to make sure that it runs through the
positioning logic, it can't rely on the new element's metrics
triggering repositioning.