Restore keyboard focus after closing a popup.
Have each window keep track of the last element that received keyboard
focus and is willing to automatically have it restored. All focusable
elements are willing by default, except for menu buttons, so that clicking
on a menu button to open a menu will still return focus to the previous
element, when that menu is closed.