Optimize scrollbar drawing.
Implement schedule_redraw() for a portion of the element's area.

explicit_redraw() passes only the requested portion to draw(), if only
a portion of the element's area was requested to be redrawn.

draw_scroll_low(), draw_slider(), and draw_scroll_high() do not draw that
portion of the scrollbar, but request its area to be redrawn, and do_draw()
checks which parts of the scrollbar are requested to be redrawn, and
only redraws those parts.