Move list peephole's metrics handling to the peephole layout manager.
In addition to the existing setting, the height_algorithm of the
peephole_style can also be a minimum and maximum number of rows.

When creating a list, if the height of the list is specified in the
number of rows, set the list_peephole_style.height_algorithm accordingly.
In all cases the the horizontal width algorithm is set to stretch_peephole.

Remove setting the list's height based on the number of rows from
update_peephole_metrics() in the peepholed_listcontainer_impl. Also
remove the code that sets the peephole's width to the peepholed list's
width, this is handled by the strech_peephole setting.

Peephole layout manager's update_or_metrics() checks the height algorithm,
and if it specifies the peephole's height based on the number of rows, it
does what update_peephole_metrics() did.