peepholed_listcontainerObj::implObj no longer needs to update_peephole_metrics
Remove the code that updates explicitly-specified dim_axis_arg for the
list's height from peepholed_listcontainerObj::implObj.

The peephole layout manager can handle this also in update_our_metrics().
Add a dim_axis_arg as an option for the height_algorithm, and make
initialize() and theme_update() responsible for recompute_vertical_metrics()
which checks if the list's height is explicitly est, and if so sets the
vertical_metrics, that update_our_metrics() installs.

Move the sanity check on minimum/maximum number of rows from
list_elementObj::implObj's constructor into new_listlayoutmanager, so the
height_value can now be private.