Implement + operation for font literals.
Implements + and += operation for fonts and theme fonts.

get_theme_font() checks if the font name includes the ';' symbol, if so
it chops it off, obtains the referenced theme font, then uses the +=
operator to modify it.

Change font::font() constructor to also check for semicolons, and do the same.

Change the "" literal to construct the default font, then use the += operator,
which is the existing "" code.