Fix replace_cell() bug.
Restoring the existing cell's parameters was overwriting (row, col),
which may be different from the ones that were stored in the grid_map.

It is not necessary to store (row, col) in thr grid_map in the first
place, so make new_grid_element_info inherit from existing_grid_element_info,
and just add the row and col members, and store existing_grid_element_info
in the grid_map.