Create label display elements using theme files.
Generate uicompiler code to recognize <element> with <label>s, and use
them to create_label().

Add new_elements and get_element() to uielements. Not very useful with
label elements per set, this is for future use.

An optional messages parameter to uigenerators links the generator to
a message catalog.

Implementation of the theme_text parameter to a text_param checks if
theme_text's generator has a message catalog, and, if so, looks up the
message in the catalog, and, to avoid "issues" will return the unicode
replacement text, otherwise theme_text converts the text label to unicode
itself. Convert the macro parsing code to use Unicode text. Add ${context:id}