Fix popup positioning bug.
Was visible in the font picker popup, the popup for the font family
combo-box.

The popup was being attached to two elements. The combo-box code was
attaching the popup to the main combo-box field. create_popup_imagebutton()
was then attaching the popup to the image button inside the combo-box field.