Add alpha channel support to the color picker.
Use a custom color_picker_current_canvas widget for the color
picker's shown color that composites the color over a checkered
background.

Add a vertical color_picker_alpha_canvas strip to the right of the
gradient square that composites the current color over a checkered
background, varying the current color's alpha channel.

Compose the color picker's gradient square over the checkered background,
so if it has an alpha channel that background will show.

Add an "A:" control, that manually adjusts the alpha channel.

Make the alpha controls optional, based on a color_picker_config setting.