Display optimizations.
Fixes display thrashing when deleting a combo-box and creating a replacement
combo-box in its place.

1) Invoke focusable containers creators before the new containers'
recalculation gets triggered.

2) The existence of a batch queue object explicitly blocks recalculation
and element update processing.

The end result is that all recalculation gets blocked until the grid
layout manager object, where the existing combo-box got removed and a new one
got created in its place, goes out of scope and gets destroyed; also the
contents of the new combo-box, if prepared in its creator, are already there.

Because of that it is no longer necessary to maintain separate queues of
batched up recalculations and which elements' visibilities were updated via
the batch queue object. That can be updated directly in the elements, and
the existence of the batch queue object will simply block their processing.