Clean-up and refactor direct pixmap scrolling.
Clean up the code for scrolling existing contents directly in the
window, instead of redrawing the entire widget from scratch. The peephole
layout manager uses this to scroll the peephole's contents, and only redraw
the newly-exposed gaps.

Need to scroll both the window's drawable, and the window_pixmap backing
drawable. This wasn't updated when the window_pixmap was implemented.

Move the copy() call from the peephole layout manager into
scroll_window_pixmap(), which scrolls both.

The peephole layout manager was checking the widget's
is_scrollable_background() and making a bunch of widget-related checks
on top of that. Move all of the relevant checks into
has_scrollable_window_pixmap().