Improve visual shading and unshading
Make shading faster by adding a window_picture to the window_pixmap,
and when a shade mcguffin gets created we will compose the shade on top
of the window_pixmap and then copy it to the window, instead of redrawing
each widget.

This requires additional logic for acquisition and destruction of shade
mcguffins, to synchronize all the race conditions.

Unshading still needs to be done by redrawing everything, but instead of
schedule_redraw_recursively() we'll do an explicit_redraw_recursively(),
this seems to be faster too.