Fix focus issue when an input field is a label_for a button.
Clicking on an input field that's a label_for a radio button transfers
the input focus to the radio button.

Replace redirect with a reference to an internal, mutable object. This
avoids having to make a copy of the entire button event.

Add another flag that logs that focus was already redirected, so we
won't do it again.

Clean up and move some logic out of the focusable_element mixin code.