Fix fragment minimum width calculation.
It should be the maximum size of all sequence of non-breaking characters
in the fragment. Except for rewrap_fragment, this is what it's used for.

So, redefine minimum_width, as such, and create an initial_width, solely
for rewrap_fragment's use.

Also cache the rich text string's direction (all left-to-right text,
all right-to-left text, or both).