Bidirection remove support.
Update removed_from_fragment's logic for adjusting positions of locations
with the removed portion of the fragment. Instead of positioning them
on the first character after the removed one, in right-to-left paragraphs
it is positions on the first character before the removed one. And update
the sanity check in remove() for right-to-left semantics.

Add merge_type parameter to merge(). The existing behavior is merge_bidi.
Specify merge_paragraph executes the merge according to the paragraph's
embedding level.

Minor cleanup, explicitly call recalculation_required().

Implement remove_info by deriving from richtext_range, and
implement range() to log which portion of each fragment gets deleted.
We use it to determine which portion of the first and last fragment
that contains the deleted range is actually deleted.

Change richtext_range to define the relevant part of the first and last
fragment to depend directly on the paragraph embedding level. Having
remove_info formally use richtext_range is now little value added,
except by leveraging complete_line().

Implement remove_at_location_no_wrap by using remove_info to determine
what to delete. Specify merge_paragraph to merge, so that for left-to-right
paragraph embedding level, the ending fragment gets merged after the
beginning fragment, as before, but for a right-to-left paragraph embedding
level it gets merged before the beginning one.