Reimplemented bidirectional insert.
richtext_insert_results also need to be updated by changes due to removing
existing text, when replace()ing the existing test, so that
remove_at_location forwards it to fragment->remove(), which updates the
richtext_insert_results accordingly.

After all the info is gather we can find where the reformatted and rewrapped
inserted text starts and ends, and reposition the cursor locations.

In create_fragments_from_inserted_text we force the embedding level
of any newlines to match the rich text's embedding level.