Implement move_for_delete().
Properly move the cursor when executing a Del operation.

When removing a range consisting of multiple ranges, after merging the
fragment with the beginning and the ending ranges we will delete the
characters which were in range before the merge, by memorizing the in-range
positions on the first and the last fragment before the merge, then
recalculating them.

Furthermore we standardize on fragment->embedding_level() as a guide as
to which part of the line is in range.

At the richtextcursorlocation level, reverse-engineer the increment that
moves the cursor left and right (since the position offsets go in the
direction of the bidi embedding level). Use that to implement
move_for_delete().