Fix bi-directional handling in hotspots.
Reimplements replace_hotspot() logic, as replace_hotspot_iterators(), moving
it into richtext_range, closer to the logic that it depends on. The
iterators are now defining the text between the hotspot start/end markers.

Need to pass a replacing_hotspots flag from replace_hotspot() into
to_canonical_order, so that it will not insert markers before/after each
hotspot.

And in insert() when it tries to figure out whether it it's doing left-to-right
or right-to-left insert, that also follows the fragment's embeddig_level(),
so everyone's on the same page.

to_canonical_order() correctly uses the default_paragraph_embedding_level().

Force LRO rendering in the date input field calendar popup.