Factor out radio button logic.
Create a radio button object class that the existing radio_image_buttonObj
inherits from and implements the mechanics of turned off or on the radio
button, with the factored out logic moved into the radio_group implementation
object that works out which radio button was turned off, and then turn it on.

radio_group now keeps track of the abstract radio button class.