Manually override direction of typed in text.
Force left-to-right or right-to-left text entry irrespective of the
direction of the typed in text.

Add two tooltip icons for visually indicating text direction override
in effect. Add the icon_images_set_element mixin to editorObj::implObj
to maintain the two icons.

Recycle the existing tooltip infrastucture to display a
separate tooltip that's manager by editorObj::implObj.

The override is checked in process_keypress, and if so it adds a bunch
of markers to the inserted text to ensure it will be forced to the given
rendering direction.

scroll_cursor_into_view calls update_attachedto_info to ensure that
the direction popup tooltips track the current cursor position.

The default implementation of direction() in elementObj::implObj returns
"automatic". editorObj::implObj overrides it and gives the direction
of the input field's text. label mixin also does it, for good measure.

Factor out the code in tooltip.C that computes the new tooltip position
into tooltip_pseudo_rectangle(), and add logic to use direction() to
determine whether the tooltip should placed to the left or the right of
the pointer position. Factor out logic that gets shared with the
editorObj::implObj for managing its own tooltip.