Do not activate shortcuts in a popup attached to a hidden widget.
It should be possible to have different widgets with popup menus that
have the same keyboard shortcut, as long as only one of them is visible.

Add an "enabled_for" parameter to enabled().

All existing calls to enable(), except for lookup_shortcut(), pass in
enabled_for::input_focus.

lookup_shortcut() passes in enabled_for::shortcut_activation.

contextmenu_popup_handlerObj overrides enabled(), and makes an additional
check, if it's for shortcut_activation then forward the call to the
atteched-to widget (in addition to handling it normally).