Implement UI for selecting a factory.
all_generators_for helper for returning a list of factories or layout
managers of a specific type.

Replace the single string parameter to create_ui() with a more generic
create_ui_info object that, in addition to the value, also gives the
handler's name.

Implement all lookup_*factory_generators types with "lookup_factory", with
the factory name being the handler name.