Automatically create referenced factories that don't exist.
When Updating generators and one of them was a new one that
references a factory: if a factory by that name doesn't exist
we'll create it.

Cleanup: store const_appgenerator_function, constant. When we're
updating them we'll clone() an existing constant object, and update
it. This turned out not to be necessary, but this is a good
change to have.

Create an appgenerator_save object, a place to stash in additional
stuff while saving a new or an updated generator.

When saving a factory value, we check if the factory does not exist,
and put a closure into the stash that will create it at the right
time during the process.