Refactor layoutmanager uicompiler.
Implement a "layout" parser, that implements common methods on the
base "layoutmanager".

Replace <creator> in a <container> with a <creator> in the "layout"
parser, which implements a <creator> in every layout manager.

Use an XML stylesheet to automate the declaration of all

<category>layout_parser() methods in uicompiler, requiring only
an explicit layout_parser() and layout_parseconfig() since the new
base "layout" parser does not have a defined category.