Split out hotspot callbacks from the text parameter.
Instead of passing text_hotspot callbacks to text_param, use identifiers
to mark start/end of hotspots in the text parameter, and pass in the
text_hotspot callbacks as an additional parameter to create_focusable_label()
and add a container to uielements. This makes it possible to fully
specify focusable labels in theme files, and supply the callbacks.